move anonymization to new file
